+++ src/main/java/com/buuz135/togetherforever/command/ForceSyncCommand.java
@@ -1,39 +1,39 @@
 package com.buuz135.togetherforever.command;
 
 import com.buuz135.togetherforever.api.IPlayerInformation;
+import com.buuz135.togetherforever.api.ISyncAction;
 import com.buuz135.togetherforever.api.ITogetherTeam;
 import com.buuz135.togetherforever.api.TogetherForeverAPI;
 import com.buuz135.togetherforever.api.command.SubCommandAction;
 import com.buuz135.togetherforever.api.data.DefaultPlayerInformation;
+import com.buuz135.togetherforever.api.data.TogetherRegistries;
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.command.PlayerNotFoundException;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
-import net.minecraft.util.text.TextFormatting;
 
-public class TeamLeaveCommand extends SubCommandAction {
+public class ForceSyncCommand extends SubCommandAction {
 
-    public TeamLeaveCommand() {
-        super("leave");
+    public ForceSyncCommand() {
+        super("forcesync");
     }
 
     @Override
     public boolean execute(MinecraftServer server, ICommandSender sender, String[] args) {
         try {
-            ITogetherTeam togetherTeam = TogetherForeverAPI.getInstance().getPlayerTeam(CommandBase.getCommandSenderAsPlayer(sender).getUniqueID());
+            EntityPlayerMP senderPlayer = CommandBase.getCommandSenderAsPlayer(sender);
+            ITogetherTeam togetherTeam = TogetherForeverAPI.getInstance().getPlayerTeam(senderPlayer.getUniqueID());
             if (togetherTeam != null) {
-                TogetherForeverAPI.getInstance().removePlayerFromTeam(togetherTeam, DefaultPlayerInformation.createInformation(CommandBase.getCommandSenderAsPlayer(sender)));
-                CommandBase.getCommandSenderAsPlayer(sender).sendMessage(new TextComponentString(TextFormatting.GREEN + "You successfully left your team."));
                 for (IPlayerInformation playerInformation : togetherTeam.getPlayers()) {
                     EntityPlayerMP playerMP = playerInformation.getPlayer();
                     if (playerMP != null) {
-                        playerMP.sendMessage(new TextComponentString(TextFormatting.RED + "Player " + sender.getName() + " left your team."));
+                        for (ISyncAction action : TogetherRegistries.getSyncActions()) {
+                            action.syncJoinPlayer(DefaultPlayerInformation.createInformation(senderPlayer), playerInformation);
+                        }
                     }
                 }
-                return true;
             }
         } catch (PlayerNotFoundException e) {
             sender.sendMessage(new TextComponentTranslation(e.getLocalizedMessage(), e.getErrorObjects()));